+++ custom_components/homematicip_local/const.py
@@ -10,7 +10,8 @@ from hahomematic.const import PLATFORMS
 from homeassistant.const import Platform
 
 DOMAIN: Final = "homematicip_local"
-HMIP_LOCAL_MIN_VERSION: Final = "2024.10.0dev0"
+HMIP_LOCAL_MIN_HA_VERSION: Final = "2024.10.0dev0"
+HMIP_LOCAL_HAHOMEMATIC_VERSION: Final = "2024.10.14"
 
 DEFAULT_DEVICE_FIRMWARE_CHECK_ENABLED: Final = True
 DEFAULT_DEVICE_FIRMWARE_CHECK_INTERVAL: Final = 21600  # 6h